college-admission-rag-agent-complete/requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# IBM watsonx.ai
ibm-watsonx-ai==1.1.0
ibm-watson-machine-learning==1.0.357

# RAG and NLP
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.18

# Document processing
pypdf2==3.0.1
python-docx==0.8.11
pandas==2.1.4

# API and utilities
requests==2.31.0
numpy==1.24.3
